test(MovieForm): use snapshot test for MovieForm component
