+++ src/components/MovieModal/MovieForm.tsx
@@ -9,8 +9,9 @@ import Input from "../Input";
 import Footer from "../Modal/Footer";
 
 import movieFormStyle from "./movie-form.module.scss";
+import { useMovieFormik } from "./utils";
 
-interface MovieFormInterface {
+export interface MovieFormInterface {
   formData: MovieDTO;
   // onChange: (value: any, fieldName: string) => any;
   submitCallback: (data: MovieDTO) => any;
@@ -23,44 +24,6 @@ function MovieForm({
   resetCallback,
 }: MovieFormInterface) {
   const genresOption = useAppSelector(selectGenres);
-  const validationSchema = yup.object().shape({
-    title: yup.string().required("Title is required!"),
-    release_date: yup
-      .date()
-      .max(new Date(), "Please select a valid release date")
-      .required("Please select a release date!"),
-    poster_path: yup
-      .string()
-      .url("Please input valid url")
-      .required("Please input preview image uri"),
-    vote_average: yup
-      .number()
-      .nullable(true)
-      .positive("Please input a positive number")
-      .max(10, "Cannot bigger than 10"),
-    genres: yup.array().min(1, "Please select at least one genre type"),
-    runtime: yup
-      .number()
-      .nullable(true)
-      .integer("Please input a integer number")
-      .positive("Please input a positive number"),
-    overview: yup
-      .string()
-      .max(500, "Cannot type more than 500 character")
-      .required("Please type overview"),
-  });
-  const formik = useFormik({
-    initialValues: formData,
-    validationSchema,
-    enableReinitialize: true,
-    onSubmit: (values) => {
-      submitCallback(values);
-    },
-    onReset: () => {
-      resetCallback();
-    },
-  });
-
   const {
     errors,
     values,
@@ -69,7 +32,11 @@ function MovieForm({
     handleReset,
     touched,
     setFieldValue,
-  } = formik;
+  } = useMovieFormik({
+    formData,
+    submitCallback,
+    resetCallback,
+  });
 
   return (
     <form
